將 GitHubUserListVC 抽出 viewModel 層

+++ ACCESS_interview/Network/NetworkManager.swift
@@ -17,7 +17,7 @@ class NetworkController: NSObject {
     return queue
   }()
   
-  func getUserList(since: Int) {
+  func getUserList(since: Int64) {
     let operation = GitHubUserListNetworkRequest(since)
     queue.addOperation(operation)
   }
